fix edit button destination
