+++ src/components/forms/fieldSetChildStyle.tsx
@@ -1,0 +1,9 @@
+export const fieldSetChildStyle = {
+    "display": "block",
+    "padding": "10px",
+    "margin": "10px auto 0",
+    "border-radius": "5px",
+    "border": "1px solid #a0a0a0",
+    "width": "400px",
+    "text-align": "center"
+};
