+++ src/components/forms/submitButtonStyle.tsx
@@ -1,0 +1,7 @@
+export const submitButtonStyle = {
+    "color": "white",
+    "background-color": "rgba(255, 0, 0, 0.5)",
+    "font-weight": "bold",
+    "font-size": "20px",
+    "cursor": "pointer"
+};
