+++ src/components/organizations/LoadedOrganization.tsx
@@ -4,12 +4,12 @@ import { headerContainerChildStyle, statementStyle, urlStyle, imageStyle } from 
 import { viewOrgStatementStyle } from "./viewOrgStatementStyle";
 import EditIcon from '@material-ui/icons/Edit';
 
-const EditButton = ({navTo, organizationURL}) => <EditIcon onClick={() => navTo(`${organizationURL}/edit`)}/>
+const EditButton = ({navTo, href}) => <EditIcon onClick={() => navTo(`${href}/edit`)}/>
 
-export const LoadedOrganization = ({ creatorEmail, title, statement, imageThumbnailURL, organizationURL, navTo }: Organization & {navTo}) => <div>
+export const LoadedOrganization = ({ creatorEmail, title, statement, imageThumbnailURL, organizationURL, navTo, href }: Organization & {navTo}) => <div>
     <div style={{ display: "inline" }}>
         <img style={imageStyle} src={imageThumbnailURL} />
-        <h1 style={headerContainerChildStyle}>{title} <EditButton {...{navTo, organizationURL}} /></h1>
+        <h1 style={headerContainerChildStyle}>{title} <EditButton {...{navTo, href}} /></h1>
     </div>
     <p style={viewOrgStatementStyle}>{statement}</p>
     <a style={urlStyle} href={organizationURL}>{organizationURL}</a>
